requirements.txt
openai
joblib
pandas
numpy
python-dotenv
Flask
requests